iommu: arm-smmu: Do not bail out if CTX fault IRQ not requested
If the context fault interrupt cannot be requested, we still
have big chances of the relative IOMMU context to just work.

While this is totally unsafe, because we wouldn't be able to
recover from a context fault on these IOMMUs, it can be
helpful on some Qualcomm platforms.

It is currently unclear whether this is a firmware bug or a
GIC driver bug.
This commit may be temporary.